Refactor PrometheusMetrics to separate telemetry module.
Moved PrometheusMetrics from lib.rs to its own telemetry module for better organization and separation of concerns.
